refactor apps.mobile_app.tasks
