+++ engine/apps/mobile_app/demo_push.py
@@ -1,4 +1,5 @@
 import json
+import logging
 import random
 import string
 import typing
@@ -6,12 +7,15 @@ import typing
 from firebase_admin.messaging import APNSPayload, Aps, ApsAlert, CriticalSound, Message
 
 from apps.mobile_app.exceptions import DeviceNotSet
-from apps.mobile_app.tasks import _construct_fcm_message, _send_push_notification, logger
 from apps.mobile_app.types import FCMMessageData, MessageType, Platform
+from apps.mobile_app.utils import construct_fcm_message, send_push_notification
 from apps.user_management.models import User
 
 if typing.TYPE_CHECKING:
     from apps.mobile_app.models import FCMDevice
+
+logger = logging.getLogger(__name__)
+logger.setLevel(logging.DEBUG)
 
 
 def send_test_push(user, critical=False):
@@ -22,11 +26,11 @@ def send_test_push(user, critical=False):
         logger.info(f"send_test_push: fcm_device not found user_id={user.id}")
         raise DeviceNotSet
     message = _get_test_escalation_fcm_message(user, device_to_notify, critical)
-    _send_push_notification(device_to_notify, message)
+    send_push_notification(device_to_notify, message)
 
 
 def _get_test_escalation_fcm_message(user: User, device_to_notify: "FCMDevice", critical: bool) -> Message:
-    # TODO: this method is copied from _get_alert_group_escalation_fcm_message
+    # TODO: this method is copied from apps.mobile_app.tasks.new_alert_group._get_fcm_message
     # to have same notification/sound/overrideDND logic. Ideally this logic should be abstracted, not repeated.
     from apps.mobile_app.models import MobileAppUserSettings
 
@@ -80,7 +84,7 @@ def _get_test_escalation_fcm_message(user: User, device_to_notify: "FCMDevice", 
         ),
     )
 
-    return _construct_fcm_message(message_type, device_to_notify, thread_id, fcm_message_data, apns_payload)
+    return construct_fcm_message(message_type, device_to_notify, thread_id, fcm_message_data, apns_payload)
 
 
 def get_test_push_title(critical: bool) -> str:
