+++ engine/apps/mobile_app/tests/tasks/test_going_oncall_notification.py
@@ -6,8 +6,17 @@ import pytest
 from django.core.cache import cache
 from django.utils import timezone
 
-from apps.mobile_app import tasks
 from apps.mobile_app.models import FCMDevice, MobileAppUserSettings
+from apps.mobile_app.tasks.going_oncall_notification import (
+    _generate_cache_key,
+    _get_fcm_message,
+    _get_notification_subtitle,
+    _get_notification_title,
+    _shift_starts_within_range,
+    _should_we_send_push_notification,
+    conditionally_send_going_oncall_push_notifications_for_all_schedules,
+    conditionally_send_going_oncall_push_notifications_for_schedule,
+)
 from apps.mobile_app.types import MessageType, Platform
 from apps.schedules.models import OnCallScheduleCalendar, OnCallScheduleICal, OnCallScheduleWeb
 from apps.schedules.models.on_call_schedule import ScheduleEvent
@@ -42,14 +51,11 @@ def _create_schedule_event(
 )
 @pytest.mark.django_db
 def test_shift_starts_within_range(timing_window_lower, timing_window_upper, seconds_until_shift_starts, expected):
-    assert (
-        tasks._shift_starts_within_range(timing_window_lower, timing_window_upper, seconds_until_shift_starts)
-        == expected
-    )
+    assert _shift_starts_within_range(timing_window_lower, timing_window_upper, seconds_until_shift_starts) == expected
 
 
 @pytest.mark.django_db
-def test_get_youre_going_oncall_notification_title(make_organization_and_user, make_user, make_schedule):
+def test_get_notification_title(make_organization_and_user, make_user, make_schedule):
     schedule_name = "asdfasdfasdfasdf"
 
     organization, user = make_organization_and_user()
@@ -95,11 +101,9 @@ def test_get_youre_going_oncall_notification_title(make_organization_and_user, m
     ##################
     # same day shift
     ##################
-    same_day_shift_title = tasks._get_youre_going_oncall_notification_title(seconds_until_going_oncall)
-    same_day_shift_subtitle = tasks._get_youre_going_oncall_notification_subtitle(schedule, same_day_shift, maus)
-    same_day_shift_no_locale_subtitle = tasks._get_youre_going_oncall_notification_subtitle(
-        schedule, same_day_shift, maus_no_locale
-    )
+    same_day_shift_title = _get_notification_title(seconds_until_going_oncall)
+    same_day_shift_subtitle = _get_notification_subtitle(schedule, same_day_shift, maus)
+    same_day_shift_no_locale_subtitle = _get_notification_subtitle(schedule, same_day_shift, maus_no_locale)
 
     assert same_day_shift_title == f"Your on-call shift starts in {humanized_time_until_going_oncall}"
     assert same_day_shift_subtitle == f"09 h 00 - 17 h 00\nSchedule {schedule_name}"
@@ -108,13 +112,9 @@ def test_get_youre_going_oncall_notification_title(make_organization_and_user, m
     ##################
     # multiple day shift
     ##################
-    multiple_day_shift_title = tasks._get_youre_going_oncall_notification_title(seconds_until_going_oncall)
-    multiple_day_shift_subtitle = tasks._get_youre_going_oncall_notification_subtitle(
-        schedule, multiple_day_shift, maus
-    )
-    multiple_day_shift_no_locale_subtitle = tasks._get_youre_going_oncall_notification_subtitle(
-        schedule, multiple_day_shift, maus_no_locale
-    )
+    multiple_day_shift_title = _get_notification_title(seconds_until_going_oncall)
+    multiple_day_shift_subtitle = _get_notification_subtitle(schedule, multiple_day_shift, maus)
+    multiple_day_shift_no_locale_subtitle = _get_notification_subtitle(schedule, multiple_day_shift, maus_no_locale)
 
     assert multiple_day_shift_title == f"Your on-call shift starts in {humanized_time_until_going_oncall}"
     assert multiple_day_shift_subtitle == f"2023-07-08 09 h 00 - 2023-07-12 17 h 00\nSchedule {schedule_name}"
@@ -132,7 +132,7 @@ def test_get_youre_going_oncall_notification_title(make_organization_and_user, m
     ],
 )
 @pytest.mark.django_db
-def test_get_youre_going_oncall_notification_subtitle(
+def test_get_notification_subtitle(
     make_organization, make_user_for_organization, make_schedule, user_timezone, expected_shift_times
 ):
     schedule_name = "asdfasdfasdfasdf"
@@ -158,28 +158,25 @@ def test_get_youre_going_oncall_notification_subtitle(
         ],
     )
 
-    assert (
-        tasks._get_youre_going_oncall_notification_subtitle(schedule, shift, maus)
-        == f"{expected_shift_times}\nSchedule {schedule_name}"
-    )
+    assert _get_notification_subtitle(schedule, shift, maus) == f"{expected_shift_times}\nSchedule {schedule_name}"
 
 
-@mock.patch("apps.mobile_app.tasks._get_youre_going_oncall_notification_subtitle")
-@mock.patch("apps.mobile_app.tasks._get_youre_going_oncall_notification_title")
-@mock.patch("apps.mobile_app.tasks._construct_fcm_message")
-@mock.patch("apps.mobile_app.tasks.APNSPayload")
-@mock.patch("apps.mobile_app.tasks.Aps")
-@mock.patch("apps.mobile_app.tasks.ApsAlert")
-@mock.patch("apps.mobile_app.tasks.CriticalSound")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification._get_notification_subtitle")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification._get_notification_title")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification.construct_fcm_message")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification.APNSPayload")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification.Aps")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification.ApsAlert")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification.CriticalSound")
 @pytest.mark.django_db
-def test_get_youre_going_oncall_fcm_message(
+def test_get_fcm_message(
     mock_critical_sound,
     mock_aps_alert,
     mock_aps,
     mock_apns_payload,
     mock_construct_fcm_message,
-    mock_get_youre_going_oncall_notification_title,
-    mock_get_youre_going_oncall_notification_subtitle,
+    mock_get_notification_title,
+    mock_get_notification_subtitle,
     make_organization,
     make_user_for_organization,
     make_schedule,
@@ -191,8 +188,8 @@ def test_get_youre_going_oncall_fcm_message(
     seconds_until_going_oncall = 600
 
     mock_construct_fcm_message.return_value = mock_fcm_message
-    mock_get_youre_going_oncall_notification_title.return_value = mock_notification_title
-    mock_get_youre_going_oncall_notification_subtitle.return_value = mock_notification_subtitle
+    mock_get_notification_title.return_value = mock_notification_title
+    mock_get_notification_subtitle.return_value = mock_notification_subtitle
 
     organization = make_organization()
     user_tz = "Europe/Amsterdam"
@@ -224,9 +221,7 @@ def test_get_youre_going_oncall_fcm_message(
         "info_notification_volume_override": json.dumps(maus.info_notification_volume_override),
     }
 
-    fcm_message = tasks._get_youre_going_oncall_fcm_message(
-        user, schedule, device, seconds_until_going_oncall, schedule_event
-    )
+    fcm_message = _get_fcm_message(user, schedule, device, seconds_until_going_oncall, schedule_event)
 
     assert fcm_message == mock_fcm_message
 
@@ -244,8 +239,8 @@ def test_get_youre_going_oncall_fcm_message(
     )
     mock_apns_payload.assert_called_once_with(aps=mock_aps.return_value)
 
-    mock_get_youre_going_oncall_notification_subtitle.assert_called_once_with(schedule, schedule_event, maus)
-    mock_get_youre_going_oncall_notification_title.assert_called_once_with(seconds_until_going_oncall)
+    mock_get_notification_subtitle.assert_called_once_with(schedule, schedule_event, maus)
+    mock_get_notification_title.assert_called_once_with(seconds_until_going_oncall)
 
     mock_construct_fcm_message.assert_called_once_with(
         MessageType.INFO, device, notification_thread_id, data, mock_apns_payload.return_value
@@ -350,7 +345,7 @@ def test_get_youre_going_oncall_fcm_message(
     ],
 )
 @pytest.mark.django_db
-def test_should_we_send_going_oncall_push_notification(
+def test_should_we_send_push_notification(
     make_organization_and_user,
     info_notifications_enabled,
     now,
@@ -366,40 +361,40 @@ def test_should_we_send_going_oncall_push_notification(
     )
 
     assert (
-        tasks.should_we_send_going_oncall_push_notification(
+        _should_we_send_push_notification(
             now, user_mobile_settings, _create_schedule_event(schedule_start, schedule_start, "12345", [])
         )
         == expected
     )
 
 
-def test_generate_going_oncall_push_notification_cache_key() -> None:
+def test_generate_cache_key() -> None:
     user_pk = "adfad"
     schedule_event = {"shift": {"pk": "dfdfdf"}}
 
     assert (
-        tasks._generate_going_oncall_push_notification_cache_key(user_pk, schedule_event)
+        _generate_cache_key(user_pk, schedule_event)
         == f"going_oncall_push_notification:{user_pk}:{schedule_event['shift']['pk']}"
     )
 
 
-@mock.patch("apps.mobile_app.tasks._send_push_notification")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification.send_push_notification")
 @pytest.mark.django_db
 def test_conditionally_send_going_oncall_push_notifications_for_schedule_schedule_not_found(
     mocked_send_push_notification,
 ):
-    tasks.conditionally_send_going_oncall_push_notifications_for_schedule(12345)
+    conditionally_send_going_oncall_push_notifications_for_schedule(12345)
     mocked_send_push_notification.assert_not_called()
 
 
-@mock.patch("apps.mobile_app.tasks.OnCallSchedule.final_events")
-@mock.patch("apps.mobile_app.tasks._send_push_notification")
-@mock.patch("apps.mobile_app.tasks.should_we_send_going_oncall_push_notification")
-@mock.patch("apps.mobile_app.tasks._get_youre_going_oncall_fcm_message")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification.OnCallSchedule.final_events")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification.send_push_notification")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification._should_we_send_push_notification")
+@mock.patch("apps.mobile_app.tasks.going_oncall_notification._get_fcm_message")
 @pytest.mark.django_db
 def test_conditionally_send_going_oncall_push_notifications_for_schedule(
-    mock_get_youre_going_oncall_fcm_message,
-    mock_should_we_send_going_oncall_push_notification,
+    mock_get_fcm_message,
+    mock_should_we_send_push_notification,
     mock_send_push_notification,
     mock_oncall_schedule_final_events,
     make_organization_and_user,
@@ -424,8 +419,8 @@ def test_conditionally_send_going_oncall_push_notifications_for_schedule(
     final_events = [schedule_event]
 
     seconds_until_shift_starts = 58989
-    mock_get_youre_going_oncall_fcm_message.return_value = mock_fcm_message
-    mock_should_we_send_going_oncall_push_notification.return_value = seconds_until_shift_starts
+    mock_get_fcm_message.return_value = mock_fcm_message
+    mock_should_we_send_push_notification.return_value = seconds_until_shift_starts
     mock_oncall_schedule_final_events.return_value = final_events
 
     schedule = make_schedule(organization, schedule_class=OnCallScheduleWeb)
@@ -434,35 +429,35 @@ def test_conditionally_send_going_oncall_push_notifications_for_schedule(
     assert cache.get(cache_key) is None
 
     # no device available
-    tasks.conditionally_send_going_oncall_push_notifications_for_schedule(schedule.pk)
+    conditionally_send_going_oncall_push_notifications_for_schedule(schedule.pk)
     mock_send_push_notification.assert_not_called()
 
     # device available
     device = FCMDevice.objects.create(user=user, registration_id="test_device_id")
     MobileAppUserSettings.objects.create(user=user, going_oncall_notification_timing=ONCALL_TIMING_PREFERENCE)
 
-    tasks.conditionally_send_going_oncall_push_notifications_for_schedule(schedule.pk)
+    conditionally_send_going_oncall_push_notifications_for_schedule(schedule.pk)
 
-    mock_get_youre_going_oncall_fcm_message.assert_called_once_with(
-        user, schedule, device, seconds_until_shift_starts, schedule_event
-    )
+    mock_get_fcm_message.assert_called_once_with(user, schedule, device, seconds_until_shift_starts, schedule_event)
     mock_send_push_notification.assert_called_once_with(device, mock_fcm_message)
     assert cache.get(cache_key) is True
 
     # we shouldn't double send the same push notification for the same user/shift
-    tasks.conditionally_send_going_oncall_push_notifications_for_schedule(schedule.pk)
+    conditionally_send_going_oncall_push_notifications_for_schedule(schedule.pk)
     assert mock_send_push_notification.call_count == 1
 
     # if the cache key expires we will resend the push notification for the same user/shift
     # (in reality we're setting a timeout on the cache key, here we will just delete it to simulate this)
     cache.delete(cache_key)
 
-    tasks.conditionally_send_going_oncall_push_notifications_for_schedule(schedule.pk)
+    conditionally_send_going_oncall_push_notifications_for_schedule(schedule.pk)
     assert mock_send_push_notification.call_count == 2
     assert cache.get(cache_key) is True
 
 
-@mock.patch("apps.mobile_app.tasks.conditionally_send_going_oncall_push_notifications_for_schedule")
+@mock.patch(
+    "apps.mobile_app.tasks.going_oncall_notification.conditionally_send_going_oncall_push_notifications_for_schedule"
+)
 @pytest.mark.django_db
 def test_conditionally_send_going_oncall_push_notifications_for_all_schedules(
     mocked_conditionally_send_going_oncall_push_notifications_for_schedule,
@@ -474,7 +469,7 @@ def test_conditionally_send_going_oncall_push_notifications_for_all_schedules(
     schedule2 = make_schedule(organization, schedule_class=OnCallScheduleICal)
     schedule3 = make_schedule(organization, schedule_class=OnCallScheduleWeb)
 
-    tasks.conditionally_send_going_oncall_push_notifications_for_all_schedules()
+    conditionally_send_going_oncall_push_notifications_for_all_schedules()
 
     mocked_conditionally_send_going_oncall_push_notifications_for_schedule.apply_async.assert_has_calls(
         [
