+++ engine/apps/mobile_app/tests/tasks/test_new_shift_swap_request.py
@@ -6,7 +6,7 @@ from django.utils import timezone
 from firebase_admin.messaging import Message
 
 from apps.mobile_app.models import FCMDevice, MobileAppUserSettings
-from apps.mobile_app.tasks import (
+from apps.mobile_app.tasks.new_shift_swap_request import (
     _get_shift_swap_requests_to_notify,
     _has_user_been_notified_for_shift_swap_request,
     _mark_shift_swap_request_notified_for_user,
@@ -102,7 +102,7 @@ def test_notify_shift_swap_requests(make_organization, make_user, make_schedule,
 
     with patch.object(notify_shift_swap_request, "delay") as mock_notify_shift_swap_request:
         with patch(
-            "apps.mobile_app.tasks._get_shift_swap_requests_to_notify",
+            "apps.mobile_app.tasks.new_shift_swap_request._get_shift_swap_requests_to_notify",
             return_value=[(ShiftSwapRequest.objects.filter(pk=shift_swap_request.pk).first(), TIMEOUT)],
         ) as mock_get_shift_swap_requests_to_notify:
             notify_shift_swap_requests()
@@ -127,7 +127,10 @@ def test_notify_shift_swap_request(make_organization, make_user, make_schedule, 
     )
 
     with patch.object(notify_user_about_shift_swap_request, "delay") as mock_notify_user_about_shift_swap_request:
-        with patch("apps.mobile_app.tasks._should_notify_user_about_shift_swap_request", return_value=True):
+        with patch(
+            "apps.mobile_app.tasks.new_shift_swap_request._should_notify_user_about_shift_swap_request",
+            return_value=True,
+        ):
             with patch.object(
                 ShiftSwapRequest,
                 "possible_benefactors",
@@ -156,7 +159,10 @@ def test_notify_shift_swap_request_should_not_notify_user(
     )
 
     with patch.object(notify_user_about_shift_swap_request, "delay") as mock_notify_user_about_shift_swap_request:
-        with patch("apps.mobile_app.tasks._should_notify_user_about_shift_swap_request", return_value=False):
+        with patch(
+            "apps.mobile_app.tasks.new_shift_swap_request._should_notify_user_about_shift_swap_request",
+            return_value=False,
+        ):
             with patch.object(
                 ShiftSwapRequest,
                 "possible_benefactors",
@@ -233,7 +239,7 @@ def test_notify_user_about_shift_swap_request(make_organization, make_user, make
         schedule, beneficiary, swap_start=swap_start, swap_end=swap_end, created_at=now
     )
 
-    with patch("apps.mobile_app.tasks._send_push_notification") as mock_send_push_notification:
+    with patch("apps.mobile_app.tasks.new_shift_swap_request.send_push_notification") as mock_send_push_notification:
         notify_user_about_shift_swap_request(shift_swap_request.pk, benefactor.pk)
 
     mock_send_push_notification.assert_called_once()
@@ -270,7 +276,7 @@ def test_notify_user_about_shift_swap_request_info_notifications_disabled(
         schedule, beneficiary, swap_start=swap_start, swap_end=swap_end, created_at=now
     )
 
-    with patch("apps.mobile_app.tasks._send_push_notification") as mock_send_push_notification:
+    with patch("apps.mobile_app.tasks.new_shift_swap_request.send_push_notification") as mock_send_push_notification:
         notify_user_about_shift_swap_request(shift_swap_request.pk, benefactor.pk)
 
     mock_send_push_notification.assert_not_called()
@@ -297,22 +303,34 @@ def test_should_notify_user(make_organization, make_user, make_schedule, make_sh
     # check _should_notify_user_about_shift_swap_request is True when info notifications are disabled
     mobile_app_settings = MobileAppUserSettings.objects.create(user=benefactor, info_notifications_enabled=False)
     with patch.object(benefactor, "is_in_working_hours", return_value=True):
-        with patch("apps.mobile_app.tasks._has_user_been_notified_for_shift_swap_request", return_value=False):
+        with patch(
+            "apps.mobile_app.tasks.new_shift_swap_request._has_user_been_notified_for_shift_swap_request",
+            return_value=False,
+        ):
             assert _should_notify_user_about_shift_swap_request(shift_swap_request, benefactor, now) is True
 
     mobile_app_settings.info_notifications_enabled = True
     mobile_app_settings.save(update_fields=["info_notifications_enabled"])
 
     with patch.object(benefactor, "is_in_working_hours", return_value=True):
-        with patch("apps.mobile_app.tasks._has_user_been_notified_for_shift_swap_request", return_value=True):
+        with patch(
+            "apps.mobile_app.tasks.new_shift_swap_request._has_user_been_notified_for_shift_swap_request",
+            return_value=True,
+        ):
             assert _should_notify_user_about_shift_swap_request(shift_swap_request, benefactor, now) is False
 
     with patch.object(benefactor, "is_in_working_hours", return_value=False):
-        with patch("apps.mobile_app.tasks._has_user_been_notified_for_shift_swap_request", return_value=False):
+        with patch(
+            "apps.mobile_app.tasks.new_shift_swap_request._has_user_been_notified_for_shift_swap_request",
+            return_value=False,
+        ):
             assert _should_notify_user_about_shift_swap_request(shift_swap_request, benefactor, now) is False
 
     with patch.object(benefactor, "is_in_working_hours", return_value=True):
-        with patch("apps.mobile_app.tasks._has_user_been_notified_for_shift_swap_request", return_value=False):
+        with patch(
+            "apps.mobile_app.tasks.new_shift_swap_request._has_user_been_notified_for_shift_swap_request",
+            return_value=False,
+        ):
             assert _should_notify_user_about_shift_swap_request(shift_swap_request, benefactor, now) is True
